refactor: convert file to typescript
